+++ dashboards-observability/public/components/visualizations/charts/bar/bar.tsx
@@ -34,7 +34,7 @@ export const Bar = ({ visualizations, layout, config }: any) => {
     : [];
   const yaxis = dataConfigTab?.metrics ? dataConfigTab.metrics.filter((item) => item.label) : [];
   const barOrientation = dataConfig?.chartStyles?.orientation || vis.orientation;
-  let isVertical = visType === visChartTypes.HorizontalBar ? barOrientation !== vis.orientation : barOrientation === vis.orientation;
+  const isVertical = visType === visChartTypes.HorizontalBar ? barOrientation !== vis.orientation : barOrientation === vis.orientation;
   let bars, valueSeries, valueForXSeries;
 
   if (!isEmpty(xaxis) && !isEmpty(yaxis)) {
